Added error / success messages

+++ src/Controllers/DevicesController.php
@@ -26,32 +26,38 @@ class DevicesController extends FrontController {
 
     public function editDevice($id){
         $device = new Device($id);
-        $device->name = Request::valueRequest("name");
-        $device->demonstration_video = Request::valueRequest("video");
-        $device->description = Request::valueRequest("description");
-        $device->location = Request::valueRequest("location");
-        $device->device_type_id = Request::valueRequest("type_id");
-        $device->flat_id = Request::valueRequest("flat_id");
-        $device->save();
-        $this->redirect(route('devices'));
+        if($device->exist()) {
+            $device->name = Request::valueRequest("name");
+            $device->demonstration_video = Request::valueRequest("video");
+            $device->description = Request::valueRequest("description");
+            $device->location = Request::valueRequest("location");
+            $device->device_type_id = Request::valueRequest("type_id");
+            $device->flat_id = Request::valueRequest("flat_id");
+            $device->save();
+            $this->redirect(route('devices'), ["message" => "Modification effectuée", "type" => "success"]);
+        } else $this->redirect(route('devices'), ["message" => "Erreur lors de la modification", "type" => "error"]);
     }
 
     public function deleteDevice($id) {
         $device = new Device($id);
-        $device->delete();
-        $this->redirect(route('devices'));
+        if($device->exist()) {
+            $device->delete();
+            $this->redirect(route('devices'), ["message" => "Suppression effectuée", "type" => "success"]);
+        } else $this->redirect(route("devices"), ["message" => "Erreur lors de la suppression", "type" => "error"]);
     }
 
     public function createDevice() {
         $device = new Device();
-        $device->name = Request::valueRequest("name");
-        $device->demonstration_video = Request::valueRequest("video");
-        $device->description = Request::valueRequest("description");
-        $device->location = Request::valueRequest("location");
-        $device->device_type_id = Request::valueRequest("type_id");
-        $device->flat_id = Request::valueRequest("flat_id");
-        $device->save();
-        $this->redirect(route('devices'));
+        if($device) { // Pas de test de duplication sur la BDD car plusieurs appareils peuvent être identiques
+            $device->name = Request::valueRequest("name");
+            $device->demonstration_video = Request::valueRequest("video");
+            $device->description = Request::valueRequest("description");
+            $device->location = Request::valueRequest("location");
+            $device->device_type_id = Request::valueRequest("type_id");
+            $device->flat_id = Request::valueRequest("flat_id");
+            $device->save();
+            $this->redirect(route('devices'), ["message" => "Ajout effectué", "type" => "success"]);
+        } else $this->redirect(route("devices"), ["message" => "Erreur lors de l'ajout", "type" => "error"]);
 
     }
 
@@ -65,7 +71,7 @@ class DevicesController extends FrontController {
             $measure->start_date = DateTime::createFromFormat("d/m/Y", $startDate)->format("Y-m-d");
             $measure->end_date = DateTime::createFromFormat("d/m/Y",    $endDate)->format("Y-m-d");
             $measure->save();
-            $this->redirect(route("device", [$id]));
+            $this->redirect(route("device", [$id]), ["message" => "Ajout effectué", "type" => "success"]);
         } else {
             $this->error_404();
         }
