+++ src/Controllers/FlatsController.php
@@ -32,22 +32,30 @@ class FlatsController extends FrontController{
 
     public function createFlat(){
         $flat = new Flat();
-        $flat->name = Request::valueRequest("name");
-        $flat->house_id = Request::valueRequest("house_id");
-        $flat->flat_type_id = Request::valueRequest("type_id");
-        $flat->security_level = Request::valueRequest("security");
-        $flat->save();
-        $this->redirect(route('flats'));
+        $test = Flat::getAll(["name" => Request::valueRequest("name")]);
+        if(empty($test) && $flat && Request::valueRequest("name") &&
+            Request::valueRequest("house_id") &&
+            Request::valueRequest("type_id") &&
+            Request::valueRequest("security")) {
+            $flat->name = Request::valueRequest("name");
+            $flat->house_id = Request::valueRequest("house_id");
+            $flat->flat_type_id = Request::valueRequest("type_id");
+            $flat->security_level = Request::valueRequest("security");
+            $flat->save();
+            $this->redirect(route('flats'), ["message" => "Appartement ajouté", "type" => "success"]);
+        } else $this->redirect(route('flats'), ["message" => "Veuillez saisir tous les champs", "type" => "error"]);
     }
 
     public function editFlat($id){
         $flat = new Flat($id);
-        $flat->name = Request::valueRequest("name");
-        $flat->house_id = Request::valueRequest("house_id");
-        $flat->flat_type_id = Request::valueRequest("type_id");
-        $flat->security_level = Request::valueRequest("security");
-        $flat->save();
-        $this->redirect(route('flats'));
+        if($flat->exist()) {
+            $flat->name = Request::valueRequest("name");
+            $flat->house_id = Request::valueRequest("house_id");
+            $flat->flat_type_id = Request::valueRequest("type_id");
+            $flat->security_level = Request::valueRequest("security");
+            $flat->save();
+            $this->redirect(route('flats'), ["message" => "Appartement modifié", "type" => "success"]);
+        } else $this->error_404();
     }
 
     public function createPage(){
@@ -56,8 +64,10 @@ class FlatsController extends FrontController{
 
     public function deleteFlat($id) {
         $flat = new Flat($id);
-        $flat->delete();
-        $this->redirect(route('flats'));
+        if($flat) {
+            $flat->delete();
+            $this->redirect(route('flats'), ["message" => "Appartement supprimé", "type" => "success"]);
+        } else $this->error_404();
     }
 
 
@@ -80,8 +90,8 @@ class FlatsController extends FrontController{
             } else {
                 SQL::insert(Flat::LODGERS_LINK_TABLE, ["user_id" => Request::valueRequest("user_id"), "flat_id" => $id, "start_date" => Request::valueRequest("sdate")]);
             }
-        }
-        $this->redirect(route('flat', [$id]));
+            $this->redirect(route('flat', [$id]), ["message" => "Habitant ajouté", "type" => "success"]);
+        } else $this->redirect(route('flat', [$id]), ["message" => "Veuillez saisir une date ainsi qu'un utilisateur valide", "type" => "error"]);
     }
 
 }
